+++ src/__tests__/handler/format.test.ts
@@ -6,7 +6,7 @@ import languages, { ProviderName } from '../../languages'
 import { disposeAll } from '../../util'
 import window from '../../window'
 import workspace from '../../workspace'
-import helper, { createTmpFile } from '../helper'
+import helper from '../helper'
 
 let nvim: Neovim
 let disposables: Disposable[] = []
@@ -40,11 +40,14 @@ describe('format handler', () => {
     })
 
     it('should throw when provider not found', async () => {
-      let doc = await workspace.document
+      await expect(async () => {
+        await commands.executeCommand('editor.action.formatDocument', 999)
+      }).rejects.toThrow(Error)
       await expect(async () => {
         await commands.executeCommand('editor.action.formatDocument')
       }).rejects.toThrow(Error)
       await expect(async () => {
+        let doc = await workspace.document
         await commands.executeCommand('editor.action.formatDocument', doc.uri)
       }).rejects.toThrow(Error)
     })
@@ -112,93 +115,6 @@ describe('format handler', () => {
     })
   })
 
-  describe('formatOnSave', () => {
-    it('should not throw when provider not found', async () => {
-      helper.updateConfiguration('coc.preferences.formatOnSaveFiletypes', ['javascript'])
-      let filepath = await createTmpFile('')
-      await helper.edit(filepath)
-      await nvim.command('setf javascript')
-      await nvim.setLine('foo')
-      await nvim.command('silent w')
-    })
-
-    it('should enable format on save', async () => {
-      helper.updateConfiguration('coc.preferences.formatOnSaveFiletypes', null)
-      helper.updateConfiguration('coc.preferences.formatOnSave', true)
-      let doc = await workspace.document
-      let res = format.shouldFormatOnSave(doc.textDocument)
-      expect(res).toBe(true)
-    })
-
-    it('should not format on save when disabled', async () => {
-      helper.updateConfiguration('coc.preferences.formatOnSaveFiletypes', ['text'])
-      disposables.push(languages.registerDocumentFormatProvider(['text'], {
-        provideDocumentFormattingEdits: document => {
-          let lines = document.getText().replace(/\n$/, '').split(/\n/)
-          let edits: TextEdit[] = []
-          for (let i = 0; i < lines.length; i++) {
-            edits.push(TextEdit.insert(Position.create(0, 0), '  '))
-          }
-          console.log(22)
-          return edits
-        }
-      }))
-      nvim.pauseNotification()
-      let filepath = await createTmpFile('a\nb\nc\n')
-      nvim.command('e ' + filepath, true)
-      nvim.command('let b:coc_disable_autoformat = 1', true)
-      nvim.command('setf text', true)
-      await nvim.resumeNotification()
-      await nvim.command('w')
-      let buf = await nvim.buffer
-      let lines = await buf.lines
-      expect(lines).toEqual(['a', 'b', 'c'])
-    })
-
-    it('should invoke format on save', async () => {
-      helper.updateConfiguration('coc.preferences.formatOnSaveFiletypes', ['text'])
-      disposables.push(languages.registerDocumentFormatProvider(['text'], {
-        provideDocumentFormattingEdits: document => {
-          let lines = document.getText().replace(/\n$/, '').split(/\n/)
-          let edits: TextEdit[] = []
-          for (let i = 0; i < lines.length; i++) {
-            let text = lines[i]
-            if (!text.startsWith(' ')) {
-              edits.push(TextEdit.insert(Position.create(i, 0), '  '))
-            }
-          }
-          return edits
-        }
-      }))
-      let filepath = await createTmpFile('a\nb\nc\n')
-      let buf = await helper.edit(filepath)
-      await nvim.command('setf text')
-      await nvim.command('w')
-      let lines = await buf.lines
-      expect(lines).toEqual(['  a', '  b', '  c'])
-    })
-
-    it('should cancel when timeout', async () => {
-      helper.updateConfiguration('coc.preferences.formatOnSaveFiletypes', ['*'])
-      let timer
-      disposables.push(languages.registerDocumentFormatProvider(['*'], {
-        provideDocumentFormattingEdits: () => {
-          return new Promise(resolve => {
-            timer = setTimeout(() => {
-              resolve(undefined)
-            }, 2000)
-          })
-        }
-      }))
-      let filepath = await createTmpFile('a\nb\nc\n')
-      await helper.edit(filepath)
-      let n = Date.now()
-      await nvim.command('w')
-      expect(Date.now() - n).toBeLessThan(1000)
-      clearTimeout(timer)
-    })
-  })
-
   describe('rangeFormat', () => {
     it('should return null when provider does not exist', async () => {
       let doc = (await workspace.document).textDocument
@@ -240,8 +156,8 @@ describe('format handler', () => {
         }
       }, 1))
       let doc = await helper.createDocument()
+      doc.setFiletype('text')
       await nvim.call('setline', [1, ['a', 'b', 'c']])
-      await nvim.command('setf text')
       await nvim.command('normal! ggvG')
       await nvim.input('<esc>')
       expect(languages.hasFormatProvider(doc.textDocument)).toBe(true)
@@ -264,9 +180,9 @@ describe('format handler', () => {
           return []
         }
       }))
-      await helper.createDocument()
+      let doc = await helper.createDocument()
+      doc.setFiletype('text')
       await nvim.call('setline', [1, ['a', 'b', 'c']])
-      await nvim.command('setf text')
       await nvim.command(`setl formatexpr=CocAction('formatSelected')`)
       await nvim.command('normal! ggvGgq')
       expect(range).toEqual({
@@ -282,9 +198,9 @@ describe('format handler', () => {
           return [TextEdit.insert(Position.create(0, 0), '  ')]
         }
       }))
-      await helper.createDocument()
+      let doc = await helper.createDocument()
+      doc.setFiletype('text')
       await nvim.setLine('foo')
-      await nvim.command('setf text')
       await helper.doAction('format')
       let line = await nvim.line
       expect(line).toEqual('  foo')
@@ -350,8 +266,8 @@ describe('format handler', () => {
           return []
         }
       }))
-      await helper.edit()
-      await nvim.command('setf text')
+      let doc = await helper.createDocument()
+      doc.setFiletype('text')
       await nvim.setLine('"')
       await nvim.input('i|')
       await helper.waitFor('getline', ['.'], '  |"end')