+++ src/__tests__/handler/inlayHint.test.ts
@@ -46,6 +46,7 @@ async function registerProvider(content: string): Promise<Disposable> {
       return hints
     }
   })
+  await helper.wait(10)
   await doc.buffer.setLines(content.split(/\n/), { start: 0, end: -1 })
   await doc.synchronize()
   return disposable
@@ -120,6 +121,7 @@ describe('InlayHint', () => {
           return null
         }
       }))
+      await helper.wait(10)
       let doc = await workspace.document
       let tokenSource = new CancellationTokenSource()
       let res = await languages.provideInlayHints(doc.textDocument, Range.create(0, 0, 3, 0), tokenSource.token)
@@ -148,6 +150,7 @@ describe('InlayHint', () => {
           return hint
         }
       }))
+      await helper.wait(10)
       let doc = await workspace.document
       let tokenSource = new CancellationTokenSource()
       let res = await languages.provideInlayHints(doc.textDocument, Range.create(0, 0, 1, 0), tokenSource.token)
@@ -174,6 +177,7 @@ describe('InlayHint', () => {
           })
         }
       }))
+      await helper.wait(10)
       let doc = await workspace.document
       let tokenSource = new CancellationTokenSource()
       let res = await languages.provideInlayHints(doc.textDocument, Range.create(0, 0, 1, 0), tokenSource.token)
@@ -470,6 +474,7 @@ describe('InlayHint', () => {
         }
       })
       disposables.push(disposable)
+      await helper.wait(10)
       let filepath = await createTmpFile('a\n\b\nc\n', disposables)
       await helper.createDocument(filepath)
       await nvim.command('setfiletype vim')