+++ history.md
@@ -2,6 +2,19 @@
 
 Notable changes of coc.nvim:
 
+## 2025-05-21
+
+- Perform format on save after execute `editor.codeActionsOnSave`, the same as
+  VSCode.
+
+## 2025-05-20
+
+- Add `winid` (current window ID) to `CursorHold` and `CursorHoldI` events handler.
+
+## 2025-05-19
+
+- Add `ModeChanged` event and `mode` property to `events`.
+
 ## 2025-05-13
 
 - Change document highlight priority to -1 to avoid override search highlight